+++ ios/public_file_saver/Sources/public_file_saver/PublicFileSaverPlugin.swift
@@ -207,4 +207,3 @@ public class PublicFileSaverPlugin: NSObject, FlutterPlugin, UIDocumentPickerDel
         return fileURL
     }
 }
-