+++ macos/public_file_saver/Package.swift
@@ -1,0 +1,21 @@
+// swift-tools-version: 5.9
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+  name: "public_file_saver",
+  platforms: [
+    .macOS("10.14")
+  ],
+  products: [
+    .library(name: "public-file-saver", targets: ["public_file_saver"])
+  ],
+  dependencies: [],
+  targets: [
+    .target(
+      name: "public_file_saver",
+      dependencies: []
+    )
+  ]
+)